+++ provider/pkg/kinds/deprecated.go
@@ -89,7 +89,6 @@ var v121 = cluster.ServerVersion{Major: 1, Minor: 21}
 var v122 = cluster.ServerVersion{Major: 1, Minor: 22}
 var v124 = cluster.ServerVersion{Major: 1, Minor: 24}
 var v125 = cluster.ServerVersion{Major: 1, Minor: 25}
-var v126 = cluster.ServerVersion{Major: 1, Minor: 26}
 var v127 = cluster.ServerVersion{Major: 1, Minor: 27}
 
 func gvkStr(gvk schema.GroupVersionKind) string {
@@ -300,7 +299,7 @@ func RemovedInVersion(gvk schema.GroupVersionKind) *cluster.ServerVersion {
 		return &v117
 	case StorageV1A1:
 		if k == CSIStorageCapacity || k == CSIStorageCapacityList {
-			return &v124
+			return &v127
 		}
 		return nil
 	default: